[RAF] expand-on-first-boot: really only do that once - reverted back
The proper solution is to wait that udev settle ends and then try to
access to the disk that in some architectures are down to a bus not
initialised at boot time because the boot happens from a separate
partition usually internal to the PCB/SoC.

After all, this script is supposed to run just one single time at the
first boot and thus the time spent on it has no impact on the every-day
performances. Moreover, in case of failure - which is more probable in
small embedded systems - the system goes in production without the full
disk capacity for which has been designed for. Small embedded systems
those usually unsupervised or remotely supervised, thus introducing a
potential source of nasty failures which could hard to detect expecially
if the disk becames overloaded or complete full.

To avoid all these risks the only reasonable approach is to set these
systems in such a way that they will reboot continously until they are
not provided with the full disk capacity. In such a way they will fail
at the day zero and this will be easily catch by the deployment tests.

This patch affects also this setting and definetely should be reverted

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>