do_bootstrap if {HOST,ROOTFS,DISTRO}_ARCH is not set warns v2
isar-debootstrap, do_bootstrap(): configuration warning added

These variables determine which debootstrap command to use because
in the general case the qemu-debootstrap should be used but it is
a slower choice when the host/target architectures match. This is
obviously managed by ISAR but the machine configuration file set
by the user could miss to properly set one of these 3 values.

So it is necessary to warn about unless the ISAR layer will properly
fix the user shortage. Also, the day the ISAR layer will correctly
address any lack, a double-check reminder/warning would be useful
because fixed things could be broken again in the future.

v.2: the bootstrap type is set coherently with the host and target
     architectures + the warning is provided whatever debootstrap
     is choosen because it is related to missing values on arches.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>